requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
mysql-connector-python
pandas
torch
transformers
